Added support for compiling mux and rfs_server under Visual Studio 2010. Patch includes move of rfs functions out of main.c in rfs_server and into rfs_transports.c/h
