+++ mux.py
@@ -1,7 +1,7 @@
 import os, sys, platform
 
 flist = "main.c"
-rfs_flist = "main.c server.c log.c deskutils.c"
+rfs_flist = "main.c server.c log.c deskutils.c rfs_transports.c"
 cdefs = "-DRFS_UDP_TRANSPORT -DRFS_INSIDE_MUX_MODE"
 socklib = ''
 ptlib = ''
