+++ rfs_server.lua
@@ -20,7 +20,7 @@ if utils.is_windows() then
     print "SIM target not supported under Windows"
     os.exit( 1 )
   end
-  flist = "main.c server.c os_io_win32.c log.c net_win32.c serial_win32.c deskutils.c"
+  flist = "main.c server.c os_io_win32.c log.c net_win32.c serial_win32.c deskutils.c rfs_transports.c"
   cdefs = cdefs .. " WIN32_BUILD"
   exeprefix = "exe"
   socklib = 'ws2_32'
