+++ rfs_server.py
@@ -13,7 +13,7 @@ if platform.system() == "Windows":
   if sim == '1':
     print "SIM target not supported under Windows"
     os.exit( 1 )
-  flist = "main.c server.c os_io_win32.c log.c net_win32.c serial_win32.c deskutils.c"
+  flist = "main.c server.c os_io_win32.c log.c net_win32.c serial_win32.c deskutils.c rfs_transports.c"
   cdefs = cdefs + " -DWIN32_BUILD"
   exeprefix = ".exe"
   socklib = '-lws2_32'
